Fix benchmark 8b f16 decomposed test
